+++ ci-profile/nodejs.sh
@@ -1,0 +1,3 @@
+export PATH="$HOME/.heroku/node/bin:$HOME/.heroku/yarn/bin:$PATH:$HOME/bin:$HOME/node_modules/.bin"
+export NODE_HOME="$HOME/.heroku/node"
+export NODE_ENV=${NODE_ENV:-test}