+++ lib/binaries.sh
@@ -1,24 +1,16 @@
-needs_resolution() {
-  local semver=$1
-  if ! [[ "$semver" =~ ^[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
-    return 0
-  else
-    return 1
-  fi
-}
-
 install_yarn() {
   local dir="$1"
-  local version="$2"
+  local version=${2:-1.x}
+  local number
+  local url
 
-  if needs_resolution "$version"; then
-    echo "Resolving yarn version ${version:-(latest)} via semver.io..."
-    local version=$(curl --silent --get --retry 5 --retry-max-time 15 --data-urlencode "range=${version}" https://semver.herokuapp.com/yarn/resolve)
+  echo "Resolving yarn version $version..."
+  if ! read number url < <(curl --silent --get --retry 5 --retry-max-time 15 --data-urlencode "range=$version" "https://nodebin.herokai.com/v1/yarn/$platform/latest.txt"); then
+    fail_bin_install yarn $version;
   fi
 
-  echo "Downloading and installing yarn ($version)..."
-  local download_url="https://yarnpkg.com/downloads/$version/yarn-v$version.tar.gz"
-  local code=$(curl "$download_url" -L --silent --fail --retry 5 --retry-max-time 15 -o /tmp/yarn.tar.gz --write-out "%{http_code}")
+  echo "Downloading and installing yarn ($number)..."
+  local code=$(curl "$url" -L --silent --fail --retry 5 --retry-max-time 15 -o /tmp/yarn.tar.gz --write-out "%{http_code}")
   if [ "$code" != "200" ]; then
     echo "Unable to download yarn: $code" && false
   fi
@@ -35,23 +27,22 @@ install_yarn() {
 }
 
 install_nodejs() {
-  local version=${1:-6.x}
+  local version=${1:-8.x}
   local dir="$2"
 
-  if needs_resolution "$version"; then
-    echo "Resolving node version $version via semver.io..."
-    local version=$(curl --silent --get --retry 5 --retry-max-time 15 --data-urlencode "range=${version}" https://semver.herokuapp.com/node/resolve)
+  echo "Resolving node version $version..."
+  if ! read number url < <(curl --silent --get --retry 5 --retry-max-time 15 --data-urlencode "range=$version" "https://nodebin.herokai.com/v1/node/$platform/latest.txt"); then
+    fail_bin_install node $version;
   fi
 
-  echo "Downloading and installing node $version..."
-  local download_url="https://s3pository.heroku.com/node/v$version/node-v$version-$os-$cpu.tar.gz"
-  local code=$(curl "$download_url" --silent --fail --retry 5 --retry-max-time 15 -o /tmp/node.tar.gz --write-out "%{http_code}")
+  echo "Downloading and installing node $number..."
+  local code=$(curl "$url" -L --silent --fail --retry 5 --retry-max-time 15 -o /tmp/node.tar.gz --write-out "%{http_code}")
   if [ "$code" != "200" ]; then
-    echo "Unable to download node $version; does it exist?" && false
+    echo "Unable to download node: $code" && false
   fi
   tar xzf /tmp/node.tar.gz -C /tmp
   rm -rf $dir/*
-  mv /tmp/node-v$version-$os-$cpu/* $dir
+  mv /tmp/node-v$number-$os-$cpu/* $dir
   chmod +x $dir/bin/*
 }
 
@@ -59,16 +50,18 @@ install_iojs() {
   local version="$1"
   local dir="$2"
 
-  if needs_resolution "$version"; then
-    echo "Resolving iojs version ${version:-(latest stable)} via semver.io..."
-    version=$(curl --silent --get  --retry 5 --retry-max-time 15 --data-urlencode "range=${version}" https://semver.herokuapp.com/iojs/resolve)
+  echo "Resolving iojs version ${version:-(latest stable)}..."
+  if ! read number url < <(curl --silent --get --retry 5 --retry-max-time 15 --data-urlencode "range=$version" "https://nodebin.herokai.com/v1/iojs/$platform/latest.txt"); then
+    fail_bin_install iojs $version;
   fi
 
-  echo "Downloading and installing iojs $version..."
-  local download_url="https://iojs.org/dist/v$version/iojs-v$version-$os-$cpu.tar.gz"
-  curl "$download_url" --silent --fail --retry 5 --retry-max-time 15 -o /tmp/node.tar.gz || (echo "Unable to download iojs $version; does it exist?" && false)
-  tar xzf /tmp/node.tar.gz -C /tmp
-  mv /tmp/iojs-v$version-$os-$cpu/* $dir
+  echo "Downloading and installing iojs $number..."
+  local code=$(curl "$url" --silent --fail --retry 5 --retry-max-time 15 -o /tmp/iojs.tar.gz --write-out "%{http_code}")
+  if [ "$code" != "200" ]; then
+    echo "Unable to download iojs: $code" && false
+  fi
+  tar xzf /tmp/iojs.tar.gz -C /tmp
+  mv /tmp/iojs-v$number-$os-$cpu/* $dir
   chmod +x $dir/bin/*
 }
 
@@ -76,24 +69,24 @@ install_npm() {
   local version="$1"
   local dir="$2"
   local npm_lock="$3"
+  local npm_version="$(npm --version)"
 
-  if $npm_lock && [ "$version" == "" ]; then
+  # If the user has not specified a version of npm, but has an npm lockfile
+  # upgrade them to npm 5.x if a suitable version was not installed with Node
+  if $npm_lock && [ "$version" == "" ] && [ "${npm_version:0:1}" -lt "5" ]; then
     echo "Detected package-lock.json: defaulting npm to version 5.x.x"
     version="5.x.x"
   fi
 
   if [ "$version" == "" ]; then
     echo "Using default npm version: `npm --version`"
+  elif [[ `npm --version` == "$version" ]]; then
+    echo "npm `npm --version` already installed with node"
   else
-    if needs_resolution "$version"; then
-      echo "Resolving npm version ${version} via semver.io..."
-      version=$(curl --silent --get --retry 5 --retry-max-time 15 --data-urlencode "range=${version}" https://semver.herokuapp.com/npm/resolve)
+    echo "Bootstrapping npm $version (replacing `npm --version`)..."
+    if ! npm install --unsafe-perm --quiet -g "npm@$version" 2>@1>/dev/null; then
+      echo "Unable to install npm $version; does it exist?" && false
     fi
-    if [[ `npm --version` == "$version" ]]; then
-      echo "npm `npm --version` already installed with node"
-    else
-      echo "Downloading and installing npm $version (replacing version `npm --version`)..."
-      npm install --unsafe-perm --quiet -g npm@$version 2>&1 >/dev/null
-    fi
+    echo "npm `npm --version` installed"
   fi
 }